+++ solutions/2024/all_solutions.h
@@ -1,9 +1,9 @@
 #pragma once
-#include "days_1_to_5/day_1.h"
-#include "days_1_to_5/day_2.h"
-#include "days_1_to_5/day_3.h"
-#include "days_1_to_5/day_4.h"
-#include "days_1_to_5/day_5.h"
-#include "days_1_to_5/day_6.h"
-#include "days_1_to_5/day_7.h"
-#include "days_1_to_5/day_8.h"
+#include "days_1_to_10/day_1.h"
+#include "days_1_to_10/day_2.h"
+#include "days_1_to_10/day_3.h"
+#include "days_1_to_10/day_4.h"
+#include "days_1_to_10/day_5.h"
+#include "days_1_to_10/day_6.h"
+#include "days_1_to_10/day_7.h"
+#include "days_1_to_10/day_8.h"
