+++ solutions/2024/days_1_to_10/day_1.cpp
